Added wall tiling to test application; added rotating of turn order based on starting player marker
