+++ azulsim/core/board/floor.py
@@ -8,7 +8,7 @@ from pydantic import ConfigDict
 from pydantic.dataclasses import dataclass
 from pydantic.types import NegativeInt, PositiveInt
 
-from ..tiles import Tile
+from ..tiles import ColoredTile, StartingPlayerMarker, Tile
 
 
 @dataclass(
